Migrate Program Errors into it's own file
