+++ src/gl/program/mod.rs
@@ -1,16 +1,16 @@
 pub mod questions;
-// mod errors;
+pub mod errors;
 // mod shaders;
 // mod programs;
 
 use std::{ffi, ptr};
-use std::fmt;
-use std::error::Error;
 
 use error::{AppResult};
 use gl;
 use gl::GlObject;
 use gl::raw::types::*;
+
+pub use gl::program::errors::*;
 
 
 /// ShaderKind is an typesafe representation of OpenGL shade types.
@@ -223,33 +223,5 @@ impl ShaderProgram {
 impl gl::GlObject for ShaderProgram {
     fn as_gl_id(&self) -> GLuint {
         self.0
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum ProgramError {
-    CompilationError(String),
-    InvalidShader(String),
-}
-
-impl fmt::Display for ProgramError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            &ProgramError::CompilationError(ref s) => write!(f, "CompilationError: {:?}", s),
-            &ProgramError::InvalidShader(ref s) => write!(f, "InvalidShader: {:?}", s),
-        }
-    }
-}
-
-impl Error for ProgramError {
-    fn description(&self) -> &str {
-        match *self {
-            ProgramError::CompilationError(_) => "Error from Compiling Shaders",
-            ProgramError::InvalidShader(_) => "Error when shader state changes from expectation",
-        }
-    }
-
-    fn cause(&self) -> Option<&Error> {
-        None
     }
 }
